requirements.txt
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
flask==3.0.3
flask-cors==4.0.1
requests==2.32.2
scikit-image==0.24.0 